makeLink function was moved to extension file
