+++ android/Buyo/app/src/main/java/com/cmpe352group4/buyo/ui/cart/CartPageFragment.kt
@@ -52,7 +52,7 @@ class CartPageFragment : BaseFragment() {
             price = 64.99,
             originalPrice = 74.99,
             brand = "Watsons",
-            vendor = Vendor(id = 12, name = "Pablos", rating = 2.43),
+            vendor = Vendor(id = "12", name = "Pablos", rating = 2.43),
             size = null,
             stockValue = null,
             category = null,
@@ -66,7 +66,7 @@ class CartPageFragment : BaseFragment() {
             price = 35.00,
             originalPrice = 70.00,
             brand = "Koton",
-            vendor = Vendor(id = 12, name = "AyseTeyze", rating = 3.21),
+            vendor = Vendor(id = "12", name = "AyseTeyze", rating = 3.21),
             size = "L",
             stockValue = null,
             category = null,
