+++ android/Buyo/app/src/main/java/com/cmpe352group4/buyo/util/extensions/GeneralExtensions.kt
@@ -2,8 +2,13 @@ package com.cmpe352group4.buyo.util.extensions
 
 import android.content.Context
 import android.os.IBinder
+import android.text.*
+import android.text.method.LinkMovementMethod
+import android.text.style.ClickableSpan
+import android.view.View
 import android.view.inputmethod.InputMethodManager
 import android.widget.ImageView
+import android.widget.TextView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
 
@@ -20,6 +25,35 @@ fun ImageView.loadFromURL(url: String) {
         .fitCenter()
         .into(this)
 }
+
+// https://stackoverflow.com/a/45727769
+fun TextView.makeLinks(vararg links: Pair<String, View.OnClickListener>) {
+    val spannableString = SpannableString(this.text)
+    for (link in links) {
+        val clickableSpan = object : ClickableSpan() {
+
+            override fun updateDrawState(textPaint: TextPaint) {
+                // use this to change the link color
+                textPaint.color = textPaint.linkColor
+                // toggle below value to enable/disable
+                // the underline shown below the clickable text
+                textPaint.isUnderlineText = true
+            }
+
+            override fun onClick(view: View) {
+                Selection.setSelection((view as TextView).text as Spannable, 0)
+                view.invalidate()
+                link.second.onClick(view)
+            }
+        }
+        val startIndexOfLink = this.text.toString().indexOf(link.first)
+        spannableString.setSpan(clickableSpan, startIndexOfLink, startIndexOfLink + link.first.length,
+            Spanned.SPAN_EXCLUSIVE_EXCLUSIVE)
+    }
+    this.movementMethod = LinkMovementMethod.getInstance()
+    this.setText(spannableString, TextView.BufferType.SPANNABLE)
+}
+
 //
 //fun ImageView.loadFromURLWithCircle(url: String) {
 //    GlideApp.with(this)
